use single db file with multiple collections, improve db loading
